docs: 담았던 정답을 수정하면 배열에 추가 되는게 아니라 수정이 되도록 함
