+++ src/pages/MBTI/MBTITest/EnergyTest.tsx
@@ -1,44 +1,14 @@
-import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components/macro';
 import Answer from './Answer';
 import { ENERGY_TEST_DB } from './constants/Test';
+import { Iprops } from './MBTITest';
 
-export interface Iprops {
-  handleSetName: (value: string, testId: number) => void;
-  onClickCheck: () => void;
-  labelName: string;
+export interface NewType extends Iprops {
   testId: number;
+  labelName: string;
 }
 
-export interface AnswerType {
-  id: number | string;
-  value: string;
-}
-
-const EnergyTest = () => {
-  const [questionID, setQuestionId] = useState<number | undefined>();
-  const [nameList, setNameList] = useState<AnswerType>({ id: '', value: '' });
-  const [isChecked, setIsChecked] = useState(false);
-  const navigate = useNavigate();
-
-  const onClickCheck = (): void => {
-    setIsChecked(!isChecked);
-  };
-
-  const handleSetName = (value: string, testId: number): void => {
-    const a = { testId, value };
-    // setNameList(nameList.concat(a));
-  };
-
-  // console.log({ nameList });
-
-  // const nextQuestion = () => {
-  //   if (nameList.length === 5) {
-  //     navigate('RelationTest');
-  //   }
-  // };
-
+const EnergyTest = ({ handleSetName, onClickCheck }: Iprops) => {
   return (
     <TestDetailContainer>
       <TestList>
